session 2 + half of session 3
1. added some new dependencies
2. added register component
3. added (dashboard component -  guards)
4. need to make a new branch for the assignments:
a. finish up authentication
b. forgot & reset password? swagger requests
c. after registering (verify) with pop-up that has 2 input elements
email is already written in the pop-up because we have it from the register component, put it in local storage then we add it to the verify pop-up
d. have a code? in register page we put the same pop-up from the previous point
e.